htpdate: really ignore timelimit when sanity checks are off
This fixes the case when a clock is started with no previous time
configured and it is therefore in epoch time: this happens for example
when a board without an RTC is bring up and needs to sync the time.
In this case, the software always refuse to synchronize because of the
timelimit limited value, which even if it's big, it's not big enough to
contemplate for the epoch time.

Signed-off-by: Angelo Compagnucci <angelo.compagnucci@gmail.com>